se_exporter/requirements.txt
prometheus_client
python-a2s
socket